RFC-167: enforce campaigns
This introduces a `checkLicense` function into the existing `campaignsEnabled` flow. We are relying on an environment variable to determine whether or not tiering should be enforced, and if set we advance to actually checking whether the current license supports the feature.

The existing permissions test cases are user-specific (e.g. is the user an admin, etc.) so the tiering tests wrap around them.

fixes https://github.com/sourcegraph/sourcegraph/issues/14029